Move arena response schemas to schemas.py; mark battle selection as placeholder
